refactor!: isolate enums and mappings
