+++ expertsystem/amplitude/canonical_decay.py
@@ -2,9 +2,11 @@
 
 from collections import OrderedDict
 
-from expertsystem.state.particle import (
+from expertsystem.nested_dicts import (
     InteractionQuantumNumberNames,
     StateQuantumNumberNames,
+)
+from expertsystem.state.particle import (
     get_interaction_property,
     get_particle_property,
 )
