+++ tests/channels/test_y_to_d0_d0bar_pi0_pi0.py
@@ -9,10 +9,8 @@ import pytest
 
 from expertsystem.amplitude.canonical_decay import CanonicalAmplitudeGenerator
 from expertsystem.amplitude.helicity_decay import HelicityAmplitudeGenerator
-from expertsystem.state.particle import (
-    InteractionQuantumNumberNames,
-    create_spin_domain,
-)
+from expertsystem.nested_dicts import InteractionQuantumNumberNames
+from expertsystem.state.particle import create_spin_domain
 from expertsystem.ui import (
     InteractionTypes,
     StateDefinition,
